further destructing the stagenav beast
