+++ client/src/components/ide/sidebar/CodeFileNav.jsx
@@ -1,0 +1,23 @@
+import React, { Component } from 'react';
+import { NavLink, Route } from 'react-router-dom';
+import SVG from 'components/SVG';
+import CodeFileOptionsNav from './CodeFileOptionsNav';
+
+class CodeFileNav extends Component {
+  render() {
+    const { basename, codeFile } = this.props;
+    const { name, id } = codeFile;
+    const path = `${basename}/file/${id}`;
+    return (
+      <li>
+        <NavLink to={path}>
+          <SVG name="codefile"/>
+          <span>{ name }</span>
+        </NavLink>
+        <Route path={path} children={({ match }) => (match && <CodeFileOptionsNav {...this.props} basename={path} codeFile={codeFile} />)} />
+      </li>
+    )
+  }
+}
+
+export default CodeFileNav;
