+++ client/src/components/ide/sidebar/StagesNav.jsx
@@ -1,12 +1,11 @@
 import React, { Component } from 'react';
-import { matchPath } from 'react-router-dom';
-import CodeFilesNav from './CodeFilesNav';
 import * as dialog from 'utils/dialog';
 import { openSidebarStage, closeSidebarStage } from 'redux/actions';
 import AddStage from './dialogs/stage/AddStage';
 import SVG from 'components/SVG';
 import './StagesNav.scss';
 import { connect } from 'react-redux';
+import StageCaret from './StageCaret';
 
 class StagesNav extends Component {
   render() {
@@ -14,7 +13,7 @@ class StagesNav extends Component {
     const sortedStages = stages.sort((a,b) => a.position - b.position);
     return (
       <ul className="stages-nav">
-        { sortedStages.map(stage => <StageNav key={stage.id} stage={stage} {...this.props} /> ) }
+        { sortedStages.map(stage => <StageCaret key={stage.id} stage={stage} {...this.props} /> ) }
         <li>
           <div className="action" onClick={() => dialog.open(AddStage, { containerId: id, position: stages.length })}>
             <SVG name="add" />
@@ -25,34 +24,6 @@ class StagesNav extends Component {
     )
   }
 }
-
-class StageNav extends Component {
-  toggle = () => {
-    const { stage: { id }, sidebarState: { stagesOpen }} = this.props;
-    if(stagesOpen.indexOf(id) >= 0) {
-      this.props.closeSidebarStage(id);
-    }
-    else {
-      this.props.openSidebarStage(id);
-    }
-  }
-  render() {
-    const { basename, location, stage: { id, title }, sidebarState: { stagesOpen }} = this.props;
-    const path = `${basename}/stage/${id}`;
-    const isOpen = stagesOpen.indexOf(id) >= 0;
-    const classes = ['directory'];
-    if(isOpen) classes.push('open');
-    const match = matchPath(location.pathname, { path });
-    if(match) classes.push('active');
-    return (
-      <li className="caret">
-        <div className={classes.join(' ')} onClick={this.toggle}> {title} </div>
-        { isOpen && <CodeFilesNav {...this.props} basename={path} /> }
-      </li>
-    )
-  }
-}
-
 
 const mapStateToProps = ({ sidebarState }) => ({ sidebarState });
 const mapDispatchToProps = { openSidebarStage, closeSidebarStage }
